refactor: improve progressView architecture and naming
- Rename LogGroups to TaskGroups for better clarity
- Move DOM formatting functions from logFormatters.js to specialized handler classes:
  - LogEntryFormatter: handles markdown formatting and special content
  - TaskGroupHeaderFormatter: handles group header creation
  - MessageTimestampExtractor: handles timestamp extraction
- Use shorter aliases (state/dom) in messageHandlers for readability
- Clean up logFormatters.js to only keep essential exports
- Update all references to use new class names and structures
