+++ src/common/modules/webviewContext.js
@@ -9,3 +9,7 @@ export function registerMessageHandlers(handlers) {
     }
   });
 }
+
+export const CHEVRON_UP_CLASS = 'codicon codicon-chevron-up';
+export const CHEVRON_DOWN_CLASS = 'codicon codicon-chevron-down';
+export const CHEVRON_RIGHT_CLASS = 'codicon codicon-chevron-right';
